+++ semcore/d3-chart/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
+## [3.0.1] - 2023-07-18
+
+### Fix
+
+- Removed ResizeObserver initiating during SSR.
 
 ## [3.0.0] - 2023-07-15